ModulatableParameter - WIP
currently throwing an unusable compile error. doesn't accept the generic definition of ModulatableParameter for some reason. first suspected that this is most likely a circular reference problem (and it might be), but after fiddling with it for an entire day, i'm beginning to suspect that maybe the implementation of ModulatableParameter had some problems that intellisense didn't catch(?)