Serialisation DAW Fix
as it turned out, the serialisation did *not* work for when a user closes and then re-opens the editor window...
had to move the SegmentableImage ownership to the AudioEngine now. which is kind of messy, but it does seem to work well at least, so yay!

also implemented a simple RegEx check for the serialisation version